return the unified form of requests belonging to `completed_nts`
In `Partial_derivation.find`, the check for inclusion in `completed_nts`
was checking that unification could be performed between the request and
the completed type, but it was not actually returning the unified form of
`req` as the previously-seen nonterminal.  This meant that `pcfg.rules`
contained productions alluding to ununified types which were not recognized
nonterminals (even though they necessarily unified with at least one
nonterminal that was in `pcfg.rules`).