correct issue with overly narrow program search
If the parameters of the grammar are too lopsided, heap search can produce
programs so long they cannot reasonably be processed by stitch.  Consequently,
it is desirable both to be able to rescale the probabilities of a DSL so that
the mean of the probabilities is the same, but the difference between the maximum
and minimum probabilities are limited to a given value, and to be able to prevent
search from returning programs about a certain size (but of any mass).