fix issue with excessive scaling
I realized that without a check that the supplied `max_diff` was not
larger than the current maximum minus minimum log likelihoods, the
rescaling operation could be increasing the differences between
log likelihoods rather than strictly putting a ceiling on them.

In a worst-case scenario, this could make a log likelihood of a program
larger than the assumed maximum value of 0.