process program during enumeration even if more than `max_programs` programs have previously been added
The previous design was not actually processing whether programs were
valid replacements after the enumeration of `max_programs` programs.

The goal is for this parameter to indicate the point after which new
representations are no longer added, not a point after which shorter
versions of programs producing existing programs can't be recorded.