fix serious bug in heap search query function
Heap search was not expanding all possible successors of programs,
rather only those in which increasing subsets of subexpressions were
replaced with their successors.  This meant that there were many
combinations of arguments that were consistently left untried.
This is likely the major reason that `add_conn`, in the previous
design, went essentially unused.  It would have been far less useful,
if not completely impotent under this restriction.